+++ scripts/constants.js
@@ -2,6 +2,31 @@
  * Contains any constants for the application
  */
 const MODULE_ID = 'sdnd-tarokka';
-export const contants = {
-    "MODULE_ID": MODULE_ID
+export const constants = {
+    "MODULE_ID": MODULE_ID,
+    "FOLDERS": {
+        "CARDS": {
+            "TAROKKA": "Tarokka Deck"
+        },
+        "MACROS": {
+            "TAROKKA": "SDND Tarokka",
+            "BTS": "Behind the Scenes"
+        }
+    },
+    "MACROS": {
+        "DEAL": "Deal Tarokka"
+    },
+    "PACKS": {
+        "COMPENDIUMS": {
+            "CARD": {
+                "TAROKKA": `${MODULE_ID}.SDND-Tarokka-Cards`
+            },
+            "MACRO": {
+                "GM": `${MODULE_ID}.SDND-Tarokka-Macros`
+            },
+            "SCENE": {
+                "TABLES": `${MODULE_ID}.SDND-Tarokka-Scenes`
+            }
+        }
+    }
 }