+++ scripts/module.js
@@ -1,5 +1,6 @@
-import { contants } from './constants.js';
+import { constants } from './constants.js';
 import { moduleSettings } from './settings.js';
+import { folders } from './folders.js';
 import { tarokka } from './tarokka.js';
 
 export let socket;
@@ -8,12 +9,12 @@ Hooks.once('init', async function() {
 });
 
 Hooks.once('socketlib.ready', async function() {
-	socket = socketlib.registerModule(contants.MODULE_ID);
+	socket = socketlib.registerModule(constants.MODULE_ID);
 });
 Hooks.once('ready', async function() {
-	//moduleSettings.registerSettings();
+	await folders.init();
+	await tarokka.init();
 	console.log("SDND Tarokka");
-	
 });
 globalThis['sdndTarokka'] = {
 	tarokka