+++ scripts/settings.js
@@ -1,4 +1,4 @@
-import { contants } from "./constants.js";
+import { constants } from "./constants.js";
 
 export let moduleSettings = {
 	'registerSettings': function _registerSettings() {
@@ -6,11 +6,11 @@ export let moduleSettings = {
 			if (!this[key].config) {
 				continue;
 			}
-			game.settings.register(contants.MODULE_ID, key, this[key].config);
+			game.settings.register(constants.MODULE_ID, key, this[key].config);
 		}
 	}
 };
 
 function getModuleSettingValue(settingName) {
-	return game?.settings?.get(contants.MODULE_ID, settingName);
+	return game?.settings?.get(constants.MODULE_ID, settingName);
 }