+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 # v1.0.0: First test release
-- Basic functionality.
+- Includes Tarokka Deck.
+- Includes two Tarokka Table Scenes.
+- Automatically imports decks and behind the scenes macro on Dealing a Deck.
+